requirements.txt

argon2-cffi==18.3.0
asn1crypto==0.24.0
astroid==2.2.5
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
distro-info==0.18
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
entrypoints==0.2.3.post3
httplib2==0.11.3
idna==2.6
isort==4.3.15
kazam==1.4.5
keyring==15.1.0
keyrings.alt==3.1
language-selector==0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
netifaces==0.10.4
olefile==0.46
pexpect==4.6.0
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.30.1
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.0.0
python-apt==1.7.0
python-debian==0.1.33
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
reportlab==3.5.6
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
six==1.12.0
systemd-python==234
typed-ast==1.3.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
virtualenv==16.4.3
wrapt==1.11.1
xkit==0.0.0
